# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentación de la API de CentralizadorBackEnd",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo de Desarrollo",
      "email": "devteam@example.com"
    },
    "description": "API REST para la gestión de datos de la plataforma CogniCare, incluyendo estudiantes, entrenadores y entrenamientos cognitivos.\n\nAutenticación\n\n **IMPORTANTE:** La mayoría de los endpoints requieren un token de autenticación (JWT Bearer).\n\nPara acceder a los endpoints protegidos, sigue estos pasos:\n1. **Obtén tu token:** Realiza una petición `POST` a `/api/login/admin` o `/api/login/entrenador` con tus credenciales.\n2. **Autoriza tus peticiones:** Haz clic en el botón **\"Authorize\"** (arriba a la derecha) y pega tu token en el campo `bearerAuth` con el formato `Bearer <TU_TOKEN_JWT>`.\n\nRoles y Permisos\n\nLa API cuenta con dos roles principales con diferentes niveles de acceso:\n\n### **Admin**\n- Gestión completa de entrenadores (CRUD).\n- Visualización de todos los informes de entrenamiento.\n- Acceso total a las utilidades del sistema.\n\n### **Entrenador**\n- Gestión de estudiantes y sus entrenamientos cognitivos.\n- Creación y seguimiento de sesiones de entrenamiento.\n- Visualización de informes de los estudiantes asignados."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido al hacer login. Formato: `Bearer <tu_token>`"
      }
    },
    "schemas": {
...
      "EstudianteEntrenamientoPorFacultad": {
        "type": "object",
        "description": "Datos resumidos de un estudiante y su entrenamiento, para listas por facultad.",
        "properties": {
          "nombre_estudiante": {
            "type": "string"
          },
          "apellido_estudiante": {
            "type": "string"
          },
          "numero_documento_estudiante": {
            "type": "string"
          },
          "tipo_documento": {
            "type": "string",
            "example": "CC"
          },
          "estado_entrenamiento": {
            "type": "string",
            "nullable": true,
            "example": "En Progreso"
          },
          "fecha_inicio_entrenamiento": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/entrenamientos-cognitivos/facultad/{facultadId}/estudiantes": {
      "get": {
        "summary": "Obtiene la lista de estudiantes y sus datos de entrenamiento para una facultad específica.",
        "tags": [
          "EntrenamientoCognitivo"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "facultadId",
            "required": true,
            "description": "ID de la facultad (UUID) para la cual se desean obtener los estudiantes.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos de estudiantes obtenidos exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Datos de estudiantes para la facultad ID <UUID> obtenidos correctamente."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EstudianteEntrenamientoPorFacultad"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "El ID de la facultad proporcionado es inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron estudiantes para la facultad especificada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "No se encontraron estudiantes para la facultad con ID <UUID>."
                    },
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/entrenamientos-cognitivos/estudiante/{siglaTipoDocEstudiante}/{numeroDocEstudiante}": {
      "patch": {
        "summary": "Modifica un entrenamiento cognitivo existente de un estudiante.",
        "tags": [
          "EntrenamientoCognitivo"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "siglaTipoDocEstudiante",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Sigla del tipo de documento del estudiante cuyo entrenamiento se va a modificar.",
            "example": "CC"
          },
          {
            "in": "path",
            "name": "numeroDocEstudiante",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Número de documento del estudiante cuyo entrenamiento se va a modificar.",
            "example": "1012345678"
          }
        ],
        "requestBody": {
          "description": "Objeto con los campos a actualizar. Solo es necesario incluir los campos que se desean cambiar.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "variablesAAgregar": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Lista de nombres de variables cognitivas a añadir al entrenamiento.",
                    "example": [
                      "Velocidad de Procesamiento"
                    ]
                  },
                  "variablesARemover": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Lista de nombres de variables cognitivas a quitar del entrenamiento.",
                    "example": [
                      "Memoria de Trabajo"
                    ]
                  },
                  "siglaTipoDocNuevoEntrenador": {
                    "type": "string",
                    "description": "Sigla del tipo de documento del nuevo entrenador. Debe ir junto con el número.",
                    "example": "CC"
                  },
                  "numeroDocNuevoEntrenador": {
                    "type": "string",
                    "description": "Número de documento del nuevo entrenador. Debe ir junto con la sigla.",
                    "example": "9876543210"
                  },
                  "nuevaFechaFinEntrenamiento": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Nueva fecha de finalización para el entrenamiento, en formato \"YYYY-MM-DD HH:MM:SS\".",
                    "example": "2026-01-15 18:00:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entrenamiento cognitivo actualizado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Entrenamiento cognitivo actualizado exitosamente."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "entrenamientoId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "detalles": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "Variable 'Memoria de Trabajo' y sus sesiones eliminadas."
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación (ej. falta un parámetro, la variable a remover no existe)."
          },
          "401": {
            "description": "No autorizado (token no válido)."
          },
          "403": {
            "description": "Prohibido (rol no permitido)."
          },
          "404": {
            "description": "Estudiante no encontrado o no tiene un entrenamiento cognitivo."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        }
      }
    },
...
    "/api/estudiantes": {
      "post": {
        "summary": "Registra un nuevo estudiante.",
        "tags": [
          "Estudiantes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "siglaTipoDocumento",
                  "nombreGenero",
                  "nombres",
                  "apellidos",
                  "numeroDocumento",
                  "fechaNacimiento",
                  "correo",
                  "programaNombres"
                ],
                "properties": {
                  "siglaTipoDocumento": {
                    "type": "string",
                    "description": "Sigla del tipo de documento (ej. \"CC\", \"TI\").",
                    "example": "CC"
                  },
                  "nombreGenero": {
                    "type": "string",
                    "description": "Nombre del género (ej. \"Masculino\", \"Femenino\", \"Otro\").",
                    "example": "Masculino"
                  },
                  "nombres": {
                    "type": "string",
                    "description": "Nombres del estudiante.",
                    "example": "Juan David"
                  },
                  "apellidos": {
                    "type": "string",
                    "description": "Apellidos del estudiante.",
                    "example": "Pérez Gómez"
                  },
                  "numeroDocumento": {
                    "type": "string",
                    "description": "Número de documento único del estudiante.",
                    "example": "1020304050"
                  },
                  "fechaNacimiento": {
                    "type": "string",
                    "format": "date",
                    "description": "Fecha de nacimiento del estudiante en formato YYYY-MM-DD.",
                    "example": "1995-08-15"
                  },
                  "correo": {
                    "type": "string",
                    "format": "email",
                    "description": "Correo electrónico único del estudiante.",
                    "example": "juan.perez@example.com"
                  },
                  "programaNombres": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Lista de nombres de programas académicos a los que se inscribe el estudiante.",
                    "example": [
                      "Ingeniería de Sistemas",
                      "Diseño Gráfico"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Estudiante registrado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Estudiante registrado exitosamente."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "description": "El UUID del estudiante recién creado."
                        },
                        "mensajeBD": {
                          "type": "string",
                          "description": "Mensaje devuelto por la función de la base de datos.",
                          "example": "Estudiante registrado exitosamente."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos o faltantes, o tipo de documento/género/programa no encontrado, o máximo de programas excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto, el estudiante ya existe (ej. correo o número de documento duplicado).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/estudiantes/{siglaTipoDoc}/{numeroDoc}": {
      "get": {
        "summary": "Obtiene la información completa de un estudiante.",
        "tags": [
          "Estudiantes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "siglaTipoDoc",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Sigla del tipo de documento del estudiante.",
            "example": "CC"
          },
          {
            "in": "path",
            "name": "numeroDoc",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Número de documento del estudiante.",
            "example": "1001004970"
          }
        ],
        "responses": {
          "200": {
            "description": "Estudiante encontrado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Estudiante encontrado exitosamente."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "nombres": {
                          "type": "string"
                        },
                        "apellidos": {
                          "type": "string"
                        },
                        "tipo_documento_sigla": {
                          "type": "string"
                        },
                        "numero_documento": {
                          "type": "string"
                        },
                        "fecha_nacimiento": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "correo": {
                          "type": "string"
                        },
                        "programas": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Estudiante no encontrado."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        }
      },
      "put": {
        "summary": "Modifica la información de un estudiante existente.",
        "tags": [
          "Estudiantes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "siglaTipoDoc",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Sigla del tipo de documento actual del estudiante.",
            "example": "CC"
          },
          {
            "in": "path",
            "name": "numeroDoc",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Número de documento actual del estudiante.",
            "example": "1020304050"
          }
        ],
        "requestBody": {
          "description": "Objeto con los campos a actualizar. Solo es necesario incluir los campos que se desean cambiar.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombres": {
                    "type": "string",
                    "description": "Los nuevos nombres del estudiante.",
                    "example": "Carlos Alberto"
                  },
                  "apellidos": {
                    "type": "string",
                    "description": "Los nuevos apellidos del estudiante.",
                    "example": "Sánchez Rodríguez"
                  },
                  "correo": {
                    "type": "string",
                    "format": "email",
                    "description": "El nuevo correo electrónico del estudiante.",
                    "example": "carlos.sanchez@example.com"
                  },
                  "fechaNacimiento": {
                    "type": "string",
                    "format": "date",
                    "description": "La nueva fecha de nacimiento en formato YYYY-MM-DD.",
                    "example": "1998-05-20"
                  },
                  "genero": {
                    "type": "string",
                    "description": "El nuevo nombre del género (ej. \"Masculino\").",
                    "example": "Masculino"
                  },
                  "programaNombres": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Reemplazo completo de la lista de programas del estudiante.",
                    "example": [
                      "Administración de Empresas"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Información del estudiante actualizada correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Información del estudiante CC-1020304050 procesada para actualización. Datos personales actualizados."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "estudianteId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "detalles": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación (ej. formato de correo inválido, programa no existe)."
          },
          "404": {
            "description": "Estudiante no encontrado."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        }
      }
    },
    "/api/estudiantes/{id}/genero": {
      "patch": {
        "summary": "Actualiza el género de un estudiante específico por su ID.",
        "tags": [
          "Estudiantes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID del estudiante (UUID)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nuevoNombreGenero"
                ],
                "properties": {
                  "nuevoNombreGenero": {
                    "type": "string",
                    "example": "Femenino"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Género actualizado exitosamente."
          },
          "400": {
            "description": "ID inválido o datos de entrada incorrectos."
          },
          "404": {
            "description": "Estudiante o género no encontrado."
          }
        }
      }
    },
...
  },
...
}